generic repo not functional cause i cant test

+++ TP4/app/src/main/java/org/Leblanc_Lauriault/tp3/DAL/AchatRepository.java
@@ -96,7 +96,7 @@ public class AchatRepository implements CRUD<Achat>
         synchronized (classe)
         {
             // set the id
-            if (a.getId() == null)
+            if ((Long)a.getId() == null)
                 a.setId(this.nextAvailableId());
 
             String serialise = gson.toJson(a);
@@ -181,6 +181,7 @@ public class AchatRepository implements CRUD<Achat>
 
     /**
      * Permet d'obtenir le prochain ID disponible pour nommer le fichier.
+     * @return
      * @return
      */
     private long nextAvailableId()
